baoming/requirements.txt
blis==0.2.4
certifi==2019.6.16
chardet==3.0.4
comtypes==1.1.7
cymem==2.0.2
Django==2.2.2
docxtpl==0.6.3
en-core-web-sm==2.0.0
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
Jinja2==2.10.1
lxml==4.3.4
MarkupSafe==1.1.1
murmurhash==1.0.2
mysqlclient==1.4.2.post1
nltk==3.4.4
numpy==1.16.4
openpyxl==2.6.2
pandas==0.24.2
Pillow==6.0.0
plac==0.9.6
preshed==2.0.1
psutil==5.6.3
python-dateutil==2.8.0
python-docx==0.8.7
pytz==2019.1
pywin32==224
requests==2.22.0
six==1.12.0
spacy==2.1.6
sqlparse==0.3.0
srsly==0.0.7
thinc==7.0.8
tqdm==4.32.2
urllib3==1.25.3
wasabi==0.2.2
xlrd==1.2.0
xlutils==2.0.0
xlwings==0.15.8
xlwt==1.3.0
